requirements.txt
tensorflow==2.12.0
numpy==1.24.3
matplotlib==3.7.2
Pillow==10.0.0
scikit-learn==1.3.0
tqdm==4.66.1
scipy==1.11.2
